Adding a, w, s, d control
Used f string where ever could have been used and used a,d,s,w for the keyboard directions. On line 19 to make it clear about x axis is for L and R and and y axis is for U and D so just changed comments a bit. Well I learned from you though ;)